Move the header, footer and metrics to a dedicated class
